fix: 🐛 In production mode, css fils couldn't load correctly
In production mode, CSS files will be concatted to a single file. After
testing, Hugo seems will skip to generate the file with same name.
And in the current design, while doing the concatenation, it will use
the same destination file name, so the custom loaded CSS files
couldn't be loaded correctly.